requirements.txt
# Core dependencies
python-dotenv==1.0.0  # For environment variable management
supabase==2.0.3       # For Supabase integration
loguru==0.7.2         # For logging

# NLP and Analysis
transformers==4.38.1   # For text analysis
torch==2.2.0          # Required by transformers
numpy==1.26.4         # For numerical operations
scikit-learn==1.4.1.post1  # For text processing and scoring
nltk==3.8.1           # For natural language processing

# Testing and Development
pytest==8.0.0         # For testing
black==24.1.1         # For code formatting
flake8==7.0.0         # For linting
